+++ core/Cargo.toml
@@ -27,7 +27,6 @@ dotenvy = "0.15.7"
 
 # utils
 futures-util = "0.3.30"
-async-trait = "0.1.81"
 bytes = "1.7.1"
 rand = "0.8.5"
 reqwest.workspace = true
@@ -36,8 +35,8 @@ reqwest.workspace = true
 base64 = "0.22.1"
 hex = "0.4.3"
 hex-literal = "0.4.1"
-serde = "1.0.204"
-serde_json = "1.0.122"
+serde.workspace = true
+serde_json.workspace = true
 
 # cli
 clap = { version = "4.5.13", features = ["derive", "env"] }
