+++ core/tests/registry_test.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "anvil")]
+
 use alloy::providers::Provider;
 use dria_oracle::{DriaOracle, DriaOracleConfig};
 use dria_oracle_contracts::OracleKind;
@@ -14,7 +16,7 @@ async fn test_registry() -> Result<()> {
         .try_init();
 
     let config = DriaOracleConfig::new_from_env()?;
-    let (node, _anvil) = DriaOracle::anvil_new(config).await?;
+    let node = DriaOracle::new(config).await?;
     assert!(node.provider.get_block_number().await? > 1);
 
     // tries to register if registered, or opposite, to trigger an error
