+++ core/tests/request_test.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "anvil")]
+
 //! Tests the request command, resulting in a task being created in the coordinator contract.
 //!
 //! 1. Requester buys some WETH, and it is approved within the request command.
@@ -28,7 +30,7 @@ async fn test_request() -> Result<()> {
 
     // node setup
     let config = DriaOracleConfig::new_from_env()?;
-    let (node, _anvil) = DriaOracle::anvil_new(config).await?;
+    let node = DriaOracle::new(config).await?;
 
     // setup account & buy some WETH
     let requester = node.connect(node.anvil_funded_wallet(None).await?);
