+++ core/tests/weth_test.rs
@@ -1,3 +1,4 @@
+#![cfg(feature = "anvil")]
 //! Using the forked blockchain, creates two accounts (alice, bob) and then,
 //!
 //! 1. Alice buys WETH
@@ -17,7 +18,7 @@ async fn test_weth_transfer() -> Result<()> {
     let amount = parse_ether("100")?;
 
     let config = DriaOracleConfig::new_from_env()?;
-    let (node, _anvil) = DriaOracle::anvil_new(config).await?;
+    let node = DriaOracle::new(config).await?;
 
     // setup alice
     let alice = node.connect(node.anvil_funded_wallet(None).await?);
